requirements.txt
ollama>=0.3.3
pyyaml>=6.0.2
schedule>=1.2.2
python-dateutil>=2.9.0
keyring>=25.0.0
